Refactor: 페이지 파일 리팩토링 (#152)
